+++ src/main/java/clover/command/AddCommand.java
@@ -1,3 +1,10 @@
+package clover.command;
+
+import clover.task.Task;
+import clover.util.TaskList;
+import clover.util.Ui;
+import clover.util.Storage;
+
 import java.io.IOException;
 
 public abstract class AddCommand extends Command {
