+++ src/main/java/clover/command/Command.java
@@ -1,3 +1,7 @@
+package clover.command;
+import clover.util.TaskList;
+import clover.util.Ui;
+import clover.util.Storage;
 import java.io.IOException;
 
 public abstract class Command {
