+++ src/main/java/clover/command/DeadlineCommand.java
@@ -1,3 +1,6 @@
+package clover.command;
+
+import clover.task.Deadline;
 public class DeadlineCommand extends AddCommand {
     public DeadlineCommand(String description, String by) {
         super(new Deadline(description, by));
