+++ src/main/java/clover/command/EventCommand.java
@@ -1,3 +1,6 @@
+package clover.command;
+
+import clover.task.Event;
 public class EventCommand extends AddCommand {
     public EventCommand(String description, String from, String to) {
         super(new Event(description, from, to));
