+++ src/main/java/clover/command/ExitCommand.java
@@ -1,3 +1,8 @@
+package clover.command;
+
+import clover.util.TaskList;
+import clover.util.Ui;
+import clover.util.Storage;
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList allTasks, Ui ui, Storage storage) {
