+++ src/main/java/clover/command/MarkCommand.java
@@ -1,17 +1,24 @@
+package clover.command;
+
+import clover.task.Task;
+import clover.util.Storage;
+import clover.util.TaskList;
+import clover.util.Ui;
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class MarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList allTasks, Ui ui, Storage storage) throws IOException {
         try {
-            Task removed = allTasks.removeTask(index);
-            ui.showDeleteTask(removed, allTasks.size());
+            Task task = allTasks.getTask(index);
+            task.markAsDone();
+            ui.showMark(task);
             storage.save(allTasks.getAllTasks());
         } catch (IndexOutOfBoundsException e) {
             ui.showError(" Oh no... This is an invalid task number!");;
