+++ src/main/java/clover/command/TodoCommand.java
@@ -1,5 +1,6 @@
-import java.io.IOException;
+package clover.command;
 
+import clover.task.Todo;
 public class TodoCommand extends AddCommand {
    public TodoCommand(String description) {
        super(new Todo(description));
