+++ src/main/java/clover/task/Deadline.java
@@ -1,3 +1,5 @@
+package clover.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 public class Deadline extends Task {
