+++ src/main/java/clover/task/Event.java
@@ -1,3 +1,5 @@
+package clover.task;
+
 public class Event extends Task {
     protected String from;
     protected String to;
