+++ src/main/java/clover/task/Task.java
@@ -1,3 +1,5 @@
+package clover.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
