+++ src/main/java/clover/task/Todo.java
@@ -1,3 +1,5 @@
+package clover.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
