+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Clover
+ Hello! I'm clover.Clover
  What can I do for you?
 ____________________________________________________________
  Got it. I've added this task:
